+++ bittytax/conv/mergers/etherscan.py
@@ -10,11 +10,17 @@ from colorama import Fore, Back
 
 from ...config import config
 from ..out_record import TransactionOutRecord
-from ..datamerge import DataMerge
+from ..datamerge import DataMerge, MergeDataRow
 from ..exceptions import UnexpectedContentError
-from ..parsers.etherscan import etherscan_txns, etherscan_tokens, get_note
+from ..parsers.etherscan import etherscan_txns, etherscan_tokens, etherscan_nfts, etherscan_int, \
+                                get_note
 
 PRECISION = Decimal('0.' + '0' * 18)
+
+TXNS = 'txn'
+TOKENS = 'token'
+NFTS = 'nft'
+INTERNAL_TXNS = 'int'
 
 def merge_etherscan(data_files):
     return do_merge_etherscan(data_files, [])
@@ -22,30 +28,42 @@ def merge_etherscan(data_files):
 def do_merge_etherscan(data_files, staking_addresses):
     merge = False
 
-    for data_row in data_files['txns'].data_rows:
-        t_tokens = find_tx_tokens(data_files['tokens'].data_rows,
-                                  data_row.row_dict['Txhash'])
+    tx_ids = {}
 
-        if t_tokens:
+    for file_id in data_files:
+        if file_id not in (TOKENS, NFTS, INTERNAL_TXNS):
+            continue
+
+        for dr in data_files[file_id].data_rows:
+            if dr.row_dict['Txhash'] in tx_ids:
+                tx_ids[dr.row_dict['Txhash']].append(MergeDataRow(dr, data_files[file_id], file_id))
+            else:
+                tx_ids[dr.row_dict['Txhash']] = [MergeDataRow(dr, data_files[file_id], file_id)]
+
+    for data_row in data_files[TXNS].data_rows:
+        t_merge = [mdr for mdr in tx_ids.get(data_row.row_dict['Txhash'], [])
+                   if not mdr.data_row.parsed]
+
+        if t_merge:
             if config.debug:
-                sys.stderr.write("%smerge: txn:  %s\n" % (Fore.GREEN, data_row))
+                sys.stderr.write("%smerge: %s:%s\n" % (Fore.GREEN, TXNS.ljust(5), data_row))
 
-            for t in t_tokens:
+            for t in t_merge:
                 if config.debug:
-                    sys.stderr.write("%smerge: token:%s\n" % (Fore.GREEN, t))
+                    sys.stderr.write("%smerge: %s:%s\n" % (Fore.GREEN, t.data_file_id.ljust(5),
+                                                           t.data_row))
 
                 data_row.parsed = True
-
         else:
             if config.debug:
                 sys.stderr.write("%smerge: txn:  %s\n" % (Fore.BLUE, data_row))
 
             continue
 
-        t_ins = [t for t in t_tokens if t.t_record and
-                 t.t_record.t_type == TransactionOutRecord.TYPE_DEPOSIT]
-        t_outs = [t for t in t_tokens if t.t_record and
-                  t.t_record.t_type == TransactionOutRecord.TYPE_WITHDRAWAL]
+        t_ins = [t.data_row for t in t_merge if t.data_row.t_record and
+                 t.data_row.t_record.t_type == TransactionOutRecord.TYPE_DEPOSIT]
+        t_outs = [t.data_row for t in t_merge if t.data_row.t_record and
+                  t.data_row.t_record.t_type == TransactionOutRecord.TYPE_WITHDRAWAL]
 
         if data_row.t_record.t_type == TransactionOutRecord.TYPE_DEPOSIT:
             t_ins.append(data_row)
@@ -56,6 +74,9 @@ def do_merge_etherscan(data_files, staking_addresses):
             output_records(data_row, t_ins, t_outs)
 
         t_ins_orig = copy.copy(t_ins)
+        fee_quantity = data_row.t_record.fee_quantity
+        fee_asset = data_row.t_record.fee_asset
+
         method_handling(t_ins, data_row, staking_addresses)
 
         # Make trades
@@ -68,16 +89,19 @@ def do_merge_etherscan(data_files, staking_addresses):
             sys.stderr.write("%sWARNING%s Merge failure for Txhash: %s\n" % (
                 Back.YELLOW+Fore.BLACK, Back.RESET+Fore.YELLOW, data_row.row_dict['Txhash']))
 
-            for t in t_tokens:
-                t.failure = UnexpectedContentError(
-                    data_files['tokens'].parser.in_header.index('Txhash'),
-                    'Txhash', t.row_dict['Txhash'])
+            for mdr in t_merge:
+                mdr.data_row.failure = UnexpectedContentError(
+                    mdr.data_file.parser.in_header.index('Txhash'),
+                    'Txhash', mdr.data_row.row_dict['Txhash'])
                 sys.stderr.write("%srow[%s] %s\n" % (
-                    Fore.YELLOW, data_files['tokens'].parser.in_header_row_num + t.line_num, t))
+                    Fore.YELLOW,
+                    mdr.data_file.parser.in_header_row_num + mdr.data_row.line_num,
+                    mdr.data_row))
+            continue
 
         # Split fees
-        t_tokens = [t for t in t_tokens if t.t_record]
-        do_fee_split(t_tokens, data_row)
+        t_all = [t for t in t_ins_orig + t_outs if t.t_record]
+        do_fee_split(t_all, data_row, fee_quantity, fee_asset)
 
         merge = True
 
@@ -85,10 +109,6 @@ def do_merge_etherscan(data_files, staking_addresses):
             output_records(data_row, t_ins_orig, t_outs)
 
     return merge
-
-def find_tx_tokens(data_rows, tx_hash):
-    return [data_row for data_row in data_rows
-            if data_row.row_dict['Txhash'] == tx_hash and not data_row.parsed]
 
 def output_records(data_row, t_ins, t_outs):
     dup = bool(data_row in t_ins + t_outs)
@@ -114,13 +134,13 @@ def method_handling(t_ins, data_row, staking_addresses):
                     t_ins.remove(staking[0])
 
                     if config.debug:
-                        sys.stderr.write("%smerge:     staking\n" % (Fore.YELLOW))
+                        sys.stderr.write("%smerge:     staking:\n" % (Fore.YELLOW))
                 else:
                     raise Exception
 
 def do_etherscan_multi_sell(t_ins, t_outs, data_row):
     if config.debug:
-        sys.stderr.write("%smerge:     trade sell(s)\n" % (Fore.YELLOW))
+        sys.stderr.write("%smerge:     trade sell(s):\n" % (Fore.YELLOW))
 
     tot_buy_quantity = 0
 
@@ -128,7 +148,7 @@ def do_etherscan_multi_sell(t_ins, t_outs, data_row):
     buy_asset = t_ins[0].t_record.buy_asset
 
     if config.debug:
-        sys.stderr.write("%smerge:     buy_quantity=%s buy_asset=%s\n" % (
+        sys.stderr.write("%smerge:       buy_quantity=%s buy_asset=%s\n" % (
             Fore.YELLOW,
             TransactionOutRecord.format_quantity(buy_quantity), buy_asset))
 
@@ -141,7 +161,7 @@ def do_etherscan_multi_sell(t_ins, t_outs, data_row):
             split_buy_quantity = buy_quantity - tot_buy_quantity
 
         if config.debug:
-            sys.stderr.write("%smerge:     split_buy_quantity=%s\n" % (
+            sys.stderr.write("%smerge:       split_buy_quantity=%s\n" % (
                 Fore.YELLOW,
                 TransactionOutRecord.format_quantity(split_buy_quantity)))
 
@@ -155,7 +175,7 @@ def do_etherscan_multi_sell(t_ins, t_outs, data_row):
 
 def do_etherscan_multi_buy(t_ins, t_outs, data_row):
     if config.debug:
-        sys.stderr.write("%smerge:     trade buy(s)\n" % (Fore.YELLOW))
+        sys.stderr.write("%smerge:     trade buy(s):\n" % (Fore.YELLOW))
 
     tot_sell_quantity = 0
 
@@ -163,7 +183,7 @@ def do_etherscan_multi_buy(t_ins, t_outs, data_row):
     sell_asset = t_outs[0].t_record.sell_asset
 
     if config.debug:
-        sys.stderr.write("%smerge:     sell_quantity=%s sell_asset=%s\n" % (
+        sys.stderr.write("%smerge:       sell_quantity=%s sell_asset=%s\n" % (
             Fore.YELLOW,
             TransactionOutRecord.format_quantity(sell_quantity), sell_asset))
 
@@ -176,7 +196,7 @@ def do_etherscan_multi_buy(t_ins, t_outs, data_row):
             split_sell_quantity = sell_quantity - tot_sell_quantity
 
         if config.debug:
-            sys.stderr.write("%smerge:     split_sell_quantity=%s\n" % (
+            sys.stderr.write("%smerge:       split_sell_quantity=%s\n" % (
                 Fore.YELLOW,
                 TransactionOutRecord.format_quantity(split_sell_quantity)))
 
@@ -188,25 +208,25 @@ def do_etherscan_multi_buy(t_ins, t_outs, data_row):
     # Remove TR for sell now it's been added to each buy
     t_outs[0].t_record = None
 
-def do_fee_split(t_tokens, data_row):
+def do_fee_split(t_all, data_row, fee_quantity, fee_asset):
     if config.debug:
-        sys.stderr.write("%smerge:     split fees\n" % (Fore.YELLOW))
+        sys.stderr.write("%smerge:     split fees:\n" % (Fore.YELLOW))
+        sys.stderr.write("%smerge:       fee_quantity=%s fee_asset=%s\n" % (
+            Fore.YELLOW,
+            TransactionOutRecord.format_quantity(fee_quantity), fee_asset))
 
     tot_fee_quantity = 0
 
-    fee_quantity = data_row.t_record.fee_quantity
-    fee_asset = data_row.t_record.fee_asset
-
-    for cnt, t in enumerate(t_tokens):
-        if cnt < len(t_tokens) - 1:
-            split_fee_quantity = (fee_quantity / len(t_tokens)).quantize(PRECISION)
+    for cnt, t in enumerate(t_all):
+        if cnt < len(t_all) - 1:
+            split_fee_quantity = (fee_quantity / len(t_all)).quantize(PRECISION)
             tot_fee_quantity += split_fee_quantity
         else:
             # Last t, use up remainder
             split_fee_quantity = fee_quantity - tot_fee_quantity if fee_quantity else None
 
         if config.debug:
-            sys.stderr.write("%smerge:     split_fee_quantity=%s\n" % (
+            sys.stderr.write("%smerge:       split_fee_quantity=%s\n" % (
                 Fore.YELLOW,
                 TransactionOutRecord.format_quantity(split_fee_quantity)))
 
@@ -215,7 +235,7 @@ def do_fee_split(t_tokens, data_row):
         t.t_record.note = get_note(data_row.row_dict)
 
     # Remove TR for fee now it's been added to each withdrawal
-    if data_row.t_record and data_row not in t_tokens:
+    if data_row.t_record and data_row not in t_all:
         if data_row.t_record.t_type == TransactionOutRecord.TYPE_SPEND:
             data_row.t_record = None
         else:
@@ -223,5 +243,8 @@ def do_fee_split(t_tokens, data_row):
             data_row.t_record.fee_asset = ''
 
 DataMerge("Etherscan fees & multi-token transactions",
-          {'txns': etherscan_txns, 'tokens': etherscan_tokens},
+          {TXNS: {'req': DataMerge.MAN, 'obj': etherscan_txns},
+           TOKENS: {'req': DataMerge.OPT, 'obj': etherscan_tokens},
+           NFTS: {'req': DataMerge.OPT, 'obj': etherscan_nfts},
+           INTERNAL_TXNS: {'req': DataMerge.OPT, 'obj': etherscan_int}},
           merge_etherscan)