+++ bittytax/conv/out_record.py
@@ -8,6 +8,8 @@ from decimal import Decimal
 from ..config import config
 from ..record import TransactionRecord
 
+STR_PRECISION = '{:0,.5f}'
+
 class TransactionOutRecord(object):
     TYPE_DEPOSIT = TransactionRecord.TYPE_DEPOSIT
     TYPE_MINING = TransactionRecord.TYPE_MINING
@@ -24,6 +26,8 @@ class TransactionOutRecord(object):
     TYPE_CHARITY_SENT = TransactionRecord.TYPE_CHARITY_SENT
     TYPE_LOST = TransactionRecord.TYPE_LOST
     TYPE_TRADE = TransactionRecord.TYPE_TRADE
+
+    ALL_TYPES = TransactionRecord.ALL_TYPES
 
     BUY_TYPES = (TYPE_DEPOSIT,
                  TYPE_MINING,
@@ -115,7 +119,7 @@ class TransactionOutRecord(object):
     def format_value(value):
         if value is not None:
             return " (%s %s)" % (
-                config.sym() + '{:0,.2f}'.format(value),
+                config.sym() + STR_PRECISION.format(value),
                 config.ccy)
         return ''